requirements.txt
google-cloud-bigquery==3.11.4
pandas>=2.1.0
numpy>=1.24.0
python-dotenv==1.0.0
